Updates to remove duplicate code between tls and verify commands.
